Bullets working near perfectly
Still must get them to only shoot on the 180 line, but it is almost complete.

When you shoot the ground, it deals no damage to the enemy and disappears.

When you shoot the enemy, it detects the enemy and deals the damage.

We will need to change the code when we inevitably add multiple enemies and enemy types